zfs: add zpool ddtload subcommand
Implement an ioctl which interfaces with a DDT adapter callback for
loading all entries of a given DDT object type, and calls it for every
DDT object that exists in a pool.

Implement the ZAP adapter callback by prefetching the entire zap object.

This subcommand enables users to pre-warm (or re-warm) the cache for DDT
entries if they reboot or otherwise perform an export/import cycle, and
skip the wait for the entries to be loaded, to restore normal I/O write
performance to the pool.

Signed-off-by:	Will Andrews <will@firepipe.net>

Conflicts:
	cmd/zpool/zpool_main.c
	include/libzfs_core.h
	include/sys/fs/zfs.h

Conflicts:
	cmd/zpool/zpool_main.c
	include/libzfs_core.h
	include/sys/fs/zfs.h
	man/man8/zpool.8