Introduce a pool property 'dedup_table_quota' to limit on-disk size of DDT
If the desired size is exceeded, new entries are no longer added,
but existing entries will still dedup.

Some changes to how the quota is both checked and calculated:

* During more experimentation, we discovered we couldn't rely
on the on-disk size of the table, since (at least with ddt-zap)
it doesn't automatically decrease.  So instead it's using the
number of entries, multiplied by ddt_entry_size().

* ddt_entry_size() is now computed (and this may be over-kill,
but it showed slightly better than a hard-coded 512), by using
the table size divided by the number of entries, and then using
the minimum of that, or 512.  With some exceptions for 0. (It
may be best to set the actual minimum to sizeof(ddt_entry_phys_t),
since it will never be smaller than that.)

* Added a new property, to be able to check the size of the
ddt.  Note that with ZFS_DEBUG, there's also another property
for debugging purposes.

* Updated the zpoolprops man page to try to explain the quota
better.

Signed-off-by: Sean Fagan <sean.fagan@klarasystems.com>
Signed-off-by: Allan Jude <allan@klarasystems.com>

Conflicts:
	include/sys/fs/zfs.h
	include/sys/spa_impl.h
	module/zcommon/zpool_prop.c
	module/zfs/ddt.c
	module/zfs/spa.c