Add the new zpool import flag --ddt-drop-unique
When the pool is imported with this flag, it skips loading all DDT ZAPs
of class UNIQUE (blocks that have not been deduped).

During the first ddt_sync() is removes the UNIQUE ZAPs entrirely

It should be used as follows:
zpool import --ddt-drop-unique <poolname>
zpool sync <poolname>
zpool export <poolname>
zpool import <poolname>

This will bulk purge the UNIQUE entries in the DDTs, restoring performance.
This operation only needs to be done as part of the upgrade process for pools
with large unique lists to make the forthcoming DDT size limiting code more
effective.

Signed-off-by: Allan Jude <allan@klarasystems.com>
Signed-off-by: Sean Fagan <sean.fagan@klarasystems.com>